+++ src/GShark/GShark.csproj
@@ -3,7 +3,7 @@
   <PropertyGroup>
     <TargetFramework>netstandard2.0</TargetFramework>
     <Title>GShark</Title>
-    <Authors>Cesare Caoduro and Mirco Bianchini</Authors>
+    <Authors>Cesare Caoduro, Mirco Bianchini, Guido Maciocci</Authors>
     <PackageLicenseExpression>MIT</PackageLicenseExpression>
     <AssemblyName>GShark</AssemblyName>
     <RootNamespace>GShark</RootNamespace>
@@ -16,7 +16,7 @@
     <PackageProjectUrl>https://gsharker.github.io/G-Shark</PackageProjectUrl>
     <PackageDescription>G-Shark is the nickname for Geometry Shark a light geometry library.</PackageDescription>
     <Summary>G-Shark library provides operation for generating and manipulating basic geometry like, curves, surfaces, meshes.</Summary>
-    <PackageTags>geometry, kernerl, csharp, nurbs, curves, bezier, polygons, intersections, mesh, solid, linear algebra, vectors</PackageTags>
+    <PackageTags>computational geometry, geometry kernel, geometry, kernel, csharp, nurbs, curves, bezier, polygons, intersections, mesh, solid, linear algebra, vectors</PackageTags>
     <Version>$(Version)</Version>
   </PropertyGroup>
